+++ R/as_polars.R
@@ -1,9 +1,7 @@
 #' To polars DataFrame
 #'
 #' [as_polars_df()] is a generic function that converts an R object to a
-#' polars DataFrame. It is basically a wrapper for [pl$DataFrame()][pl_DataFrame],
-#' but has special implementations for Apache Arrow-based objects such as
-#' polars [LazyFrame][LazyFrame_class] and [arrow::Table].
+#' [polars DataFrame][DataFrame_class].
 #'
 #' For [LazyFrame][LazyFrame_class] objects, this function is a shortcut for
 #' [$collect()][LazyFrame_collect] or [$fetch()][LazyFrame_fetch], depending on
@@ -67,17 +65,35 @@ as_polars_df.default = function(x, ...) {
 #' @param make_names_unique A logical flag to replace duplicated column names
 #' with unique names. If `FALSE` and there are duplicated column names, an
 #' error is thrown.
+#' @inheritParams as_polars_df.ArrowTabular
 #' @export
-as_polars_df.data.frame = function(x, ..., rownames = NULL, make_names_unique = TRUE) {
-  if ((anyDuplicated(names(x)) > 0) && make_names_unique) {
-    names(x) = make.unique(names(x), sep = "_")
+as_polars_df.data.frame = function(
+    x,
+    ...,
+    rownames = NULL,
+    make_names_unique = TRUE,
+    schema = NULL,
+    schema_overrides = NULL) {
+  uw = \(res) unwrap(res, "in as_polars_df():")
+
+  if (anyDuplicated(names(x)) > 0) {
+    col_names_orig = names(x)
+    if (make_names_unique) {
+      names(x) = make.unique(col_names_orig, sep = "_")
+    } else {
+      Err_plain(
+        paste(
+          "conflicting column names not allowed:",
+          paste(unique(col_names_orig[duplicated(col_names_orig)]), collapse = ", ")
+        )
+      ) |>
+        uw()
+    }
   }
 
   if (is.null(rownames)) {
-    pl$DataFrame(x, make_names_unique = FALSE)
+    df_to_rpldf(x, schema = schema, schema_overrides = schema_overrides)
   } else {
-    uw = \(res) unwrap(res, "in as_polars_df():")
-
     if (length(rownames) != 1L || !is.character(rownames) || is.na(rownames)) {
       Err_plain("`rownames` must be a single string, or `NULL`") |>
         uw()
@@ -102,7 +118,7 @@ as_polars_df.data.frame = function(x, ..., rownames = NULL, make_names_unique = 
 
     pl$concat(
       pl$Series(old_rownames, name = rownames),
-      pl$DataFrame(x, make_names_unique = FALSE),
+      df_to_rpldf(x, schema = schema, schema_overrides = schema_overrides),
       how = "horizontal"
     )
   }
@@ -133,7 +149,7 @@ as_polars_df.RPolarsDynamicGroupBy = as_polars_df.RPolarsGroupBy
 #' @rdname as_polars_df
 #' @export
 as_polars_df.RPolarsSeries = function(x, ...) {
-  pl$DataFrame(x)
+  pl$select(x)
 }
 
 
@@ -197,7 +213,7 @@ as_polars_df.ArrowTabular = function(
     rechunk = TRUE,
     schema = NULL,
     schema_overrides = NULL) {
-  arrow_to_rdf(
+  arrow_to_rpldf(
     x,
     rechunk = rechunk,
     schema = schema,
@@ -236,7 +252,7 @@ as_polars_df.nanoarrow_array_stream = function(x, ...) {
       }
     }
 
-    out = do.call(pl$DataFrame, data_cols)
+    out = do.call(pl$select, data_cols)
   } else {
     out = pl$DataFrame() # TODO: support creating 0-row DataFrame
   }